server: move linux specific functions to sockopt_linux.go
